device: consider unrealized devices as managed
A unrealized device represents a "potential" software device, that
doesn't exist in kernel at the moment but will exist if a connection
gets activated.

Until now, nm_device_get_managed() returned FALSE for unrealized
devices. Since the function is primarily used to decide whether the
device has flags that make it unmanaged, it seems more correct to
consider them as managed.

This change doesn't affect the "managed" D-Bus property of the Device
object, as the property value is determined only by the device state.